Refactor ContactForm and Greeting components
Optimized the code by shifting the contact form code into a separate ContactForm component and creating a separate ContactMeButton component in the Greeting.tsx file. This made the 'Contact Me' button and ContactForm modular and reusable. Furthermore, adjusted the visibility condition of the 'Contact Me' button, which now appears regardless of the 'isExpanded' state.